fix: Create separate providers for each DynamoDB seeder
The `CustomResourceProvider.getOrCreate()` method returns a stack-level
singleton ARN (service token) for the custom resource provider.
Thats why it expects the second
parameter to be a unique ID, for which to either get a pre-existing
provider or create a new one.
The specified `policyStatements` will only be applied in the first call,
i.e. when the provider is created, and not for stacked for any
subsequent calls, i.e. when the pre-existing provider is reused.